+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Random grading now assigns the same grade to all parts. Added a config which can revert to the old behavior.
+### Fixed
+- Some issues with the JEI plugin on servers
+
 ## [1.2.0] - 2019-07-01
 Updated for Minecraft 1.14.3